Changed _deleted semantics to only mark records as deleted
- Improved the tests so they can run towards a PostgreSQL table instead of SQLite also.
- Added different semantics based on whether `id` is present in the record or not. (The code is intrinsically growing more and more `ember-pouch`-centric.)
