+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2017-09-26 16:04:24 +0300 using RuboCop version 0.50.0.
+# on 2017-10-02 12:02:20 +0300 using RuboCop version 0.50.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -10,10 +10,10 @@
 Metrics/AbcSize:
   Max: 36
 
-# Offense count: 3
+# Offense count: 1
 # Configuration parameters: CountComments, ExcludedMethods.
 Metrics/BlockLength:
-  Max: 54
+  Max: 59
 
 # Offense count: 1
 # Configuration parameters: CountBlocks.
@@ -23,7 +23,7 @@ Metrics/BlockNesting:
 # Offense count: 2
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 211
+  Max: 202
 
 # Offense count: 2
 Metrics/CyclomaticComplexity:
@@ -32,13 +32,8 @@ Metrics/CyclomaticComplexity:
 # Offense count: 11
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
-  Max: 30
+  Max: 31
 
 # Offense count: 1
 Metrics/PerceivedComplexity:
   Max: 10
-
-# Offense count: 1
-Style/MultipleComparison:
-  Exclude:
-    - 'lib/couchdb_to_sql/builders/table.rb'
