+++ Rakefile
@@ -11,9 +11,14 @@ RuboCop::RakeTask.new
 Rake::TestTask.new do |t|
   t.libs << 'test'
   t.test_files = FileList.new('test/unit/**/*.rb')
+
+  # Without this setting, the unit test running generates a load of warnings in unrelated/3rd party gems, which obscures the
+  # real output of the test runs and makes it harder to read.
+  t.warning = false
 end
 
-# The tests are unfortunately at the moment MRI only, because of an Sqlite dependency: https://github.com/ecraft/couch_tap/issues/9
+# The tests are unfortunately at the moment MRI only, because of an Sqlite dependency:
+# https://github.com/ecraft/couchdb_to_sql/issues/9
 if defined?(JRUBY_VERSION)
   desc 'Runs Rubocop linting'
   task default: :rubocop
