+++ couchdb_to_sql.gemspec
@@ -14,7 +14,8 @@ Gem::Specification.new do |s|
 
   s.files         = `git ls-files`.split("\n")
   s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
-  s.executables   = `git ls-files -- bin/*`.split("\n").map { |f| File.basename(f) }
+  s.bindir        = 'exe'
+  s.executables   = `git ls-files -- exe/*`.split("\n").map { |f| File.basename(f) }
   s.require_paths = ['lib']
 
   s.add_dependency 'activesupport', '~> 5.0'
@@ -26,5 +27,6 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'mocha'
   s.add_development_dependency 'rake', '~> 12.0'
   s.add_development_dependency 'rubocop'
+  s.add_development_dependency 'simplecov', '~> 0.15'
   s.add_development_dependency 'test-unit', '~> 3.2'
 end
