--- lib/couchdb_to_sql/destroyers/collection.rb
@@ -1,32 +1,0 @@
-
-# frozen_string_literal: true
-
-module CouchdbToSql
-  module Destroyers
-    #
-    # Collection Destroyer. Go through each sub-table definition and remove
-    # all references to the parent document.
-    #
-    class Collection
-      attr_reader :parent
-
-      def initialize(parent, _opts = {}, &block)
-        @_tables = []
-        @parent  = parent
-
-        instance_eval(&block)
-      end
-
-      def execute
-        # Just go through each table and ask it to execute itself
-        @_tables.each(&:execute)
-      end
-
-      #### DSL Methods
-
-      def table(name, opts = {}, &block)
-        @_tables << Table.new(parent, name, opts, &block)
-      end
-    end
-  end
-end
