--- test/unit/destroyers/collection_test.rb
@@ -1,55 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-module Destroyers
-  class CollectionTest < Test::Unit::TestCase
-    def setup
-      @parent = mock
-    end
-
-    def test_initialize_collection
-      @collection = CouchdbToSql::Destroyers::Collection.new(@parent) do
-        # nothing
-      end
-      assert_equal @collection.parent, @parent
-    end
-
-    def test_raise_error_if_no_block
-      assert_raise ArgumentError do
-        @collection = CouchdbToSql::Destroyers::Collection.new(@parent)
-      end
-    end
-
-    def test_defining_table
-      @table = mock
-      CouchdbToSql::Destroyers::Table.expects(:new).with(@parent, :invoice_items, {}).returns(@table)
-      @collection = CouchdbToSql::Destroyers::Collection.new(@parent) do
-        table :invoice_items
-      end
-      tables = @collection.instance_eval('@_tables')
-      assert_equal tables.length, 1
-      assert_equal tables.first, @table
-    end
-
-    def test_defining_tables
-      CouchdbToSql::Destroyers::Table.expects(:new).twice
-      @collection = CouchdbToSql::Destroyers::Collection.new(@parent) do
-        table :invoice_items
-        table :invoice_entries
-      end
-      tables = @collection.instance_eval('@_tables')
-      assert_equal tables.length, 2
-    end
-
-    def test_execution
-      @table = mock
-      CouchdbToSql::Destroyers::Table.expects(:new).returns(@table)
-      @collection = CouchdbToSql::Destroyers::Collection.new(@parent) do
-        table :invoice_items
-      end
-      @table.expects(:execute)
-      @collection.execute
-    end
-  end
-end
